add answer hw_grilya_2025_01_22

+++ group/homework/grilya/hw_2025_01_22_part1.py
@@ -1,7 +1,7 @@
 
 print('''Пользователь ебет в рот этот синтаксиз и с пятой подсказки вспоминает про ебучие кавычки, введите цифры и нажмите энтер''')
-biba: float = input()
-boba: float = input()
+biba: int = input()
+boba: int = input()
 x: float = float(biba)
 y: float = float(boba)
 print(f'''{x} + {y} = {x + y}''')
